requirements.txt
pydantic-ai>=0.1.4
chainlit==0.7.700
pydantic>=2.5.2
langchain-core>=0.1.17
langgraph>=0.0.24
tavily-python>=0.3.1
openai>=1.3.7
python-dotenv>=1.0.0
typing-extensions>=4.9.0
langsmith>=0.0.83  
python-multipart>=0.0.6  
tiktoken>=0.5.2  
